feat(workflows): implement two-tier grading and refactor benchmark runner into modules
